refactor: move publish related from utils to publish
