+++ providers/ic/src/publish/doc.publish.ts
@@ -1,14 +1,14 @@
 import {Doc, DocData, docPublishData, DocPublishData, PublishData} from '@deckdeckgo/editor';
 import {EnvStore} from '../stores/env.store';
 import {setData} from '../api/data.api';
-import {uploadSocialImage} from './publish.social.utils';
+import {uploadSocialImage} from './social.publish';
 import {
   initIndexHTML,
   initUpload,
   StorageUpload,
   updateMetaData,
   uploadPublishFileIC
-} from './publish.utils';
+} from './common.publish';
 
 export const publishDoc = async ({
   doc: docSource,
