+++ providers/ic/src/publish/resources.publish.ts
@@ -5,10 +5,10 @@ import {
   _SERVICE as StorageBucketActor
 } from '../canisters/storage/storage.did';
 import {EnvStore} from '../stores/env.store';
-import {toNullable} from './did.utils';
-import {BucketActor} from './manager.utils';
-import {getAuthor} from './publish.utils';
-import {getStorageActor, upload} from './storage.utils';
+import {toNullable} from '../utils/did.utils';
+import {BucketActor} from '../utils/manager.utils';
+import {getAuthor} from './common.publish';
+import {getStorageActor, upload} from '../utils/storage.utils';
 
 type KitMimeType = 'text/javascript' | 'text/plain' | 'application/manifest+json' | 'text/css';
