+++ providers/ic/src/publish/rss.publish.ts
